Split parsables into individual files
