+++ docs/parsables/color.md
@@ -1,0 +1,9 @@
+`color`
+
+Can also be parsed directly from a list `[r,g,b]`, or a number (`0xRRGGBB`)
+
+| Value | Type   | Description     |
+|-------|--------|-----------------|
+| `r`   | number | Red component   |
+| `g`   | number | Green component |
+| `b`   | number | Blue component  |
