+++ docs/parsables/commands/message-context-menu-builder.md
@@ -1,0 +1,5 @@
+`message_context_menu_builder`
+
+| Value         | Type   | Description      |
+|---------------|--------|------------------|
+| `name`        | String | The command name |
