+++ docs/parsables/commands/slash-command-builder.md
@@ -1,0 +1,7 @@
+`slash_command_builder`
+
+| Value         | Type                          | Description                    |
+|---------------|-------------------------------|--------------------------------|
+| `name`        | String                        | The slash command name         |
+| `description` | String                        | The description                |
+| `options`     | List of slash command options | The sub-options of the command |
