+++ docs/parsables/commands/slash-command-option-choice.md
@@ -1,0 +1,6 @@
+`slash_command_option_choice`
+
+| Value   | Type   | Description                                                                    |
+|---------|--------|--------------------------------------------------------------------------------|
+| `name`  | String | The visible autocompleted filled in choice for the option                      |
+| `value` | String | The value that will be received in the slash command event as the option value |
