+++ docs/parsables/commands/user-context-menu-builder.md
@@ -1,0 +1,5 @@
+`user_context_menu_builder`
+
+| Value         | Type   | Description      |
+|---------------|--------|------------------|
+| `name`        | String | The command name |
