+++ docs/parsables/components/component.md
@@ -1,0 +1,8 @@
+`component`
+
+Can be either a [Button](#Button), a [Select menu](#Select menu) or a [Text input](#Text-input).
+In all cases, the values from the corresponding parsables must be included.
+
+| Value       | Type   | Description                                     |
+|-------------|--------|-------------------------------------------------|
+| `component` | String | Must be `button`, `select_menu` or `text_input` |
