+++ docs/parsables/components/modal.md
@@ -1,0 +1,7 @@
+`modal`
+
+| Value              | Type                                             | Description                                                                                                                                      |
+|--------------------|--------------------------------------------------|--------------------------------------------------------------------------------------------------------------------------------------------------|
+| `id`               | String                                           | Custom id for this modal. Used to identify when submitted                                                                                        |
+| `title`            | String                                           | Title of the modal popup                                                                                                                         |
+| `components`       | List of List of [Message components](#Component) | Each item in this list is one row of message components, and each sub-list (row) contains Components (Currently only [Text input](#Text-input)s) |
