+++ docs/parsables/components/select-menu-option.md
@@ -1,0 +1,9 @@
+`select_menu_option`
+
+| Value            | Type                             | Description                                |
+|------------------|----------------------------------|--------------------------------------------|
+| `value`          | String                           | The internal value of this option          |
+| `label`          | String                           | The text shown on the option               |
+| `emoji`          | String or Emoji value (optional) | The emoji shown on the option              |
+| `description`    | String (optional)                | A description for this option              |
+| `default_option` | boolean (optional)               | Whether this is selected by default or not |
