+++ docs/parsables/components/text-input.md
@@ -1,0 +1,12 @@
+`text_input`
+
+| Value         | Type               | Description                                                                                                                      |
+|---------------|--------------------|----------------------------------------------------------------------------------------------------------------------------------|
+| `id`          | String             | The id of the text input                                                                                                         |
+| `style`       | String             | The [style](https://discord.com/developers/docs/interactions/message-components#text-inputs-text-input-styles) of the text input |
+| `label`       | String             | The label of the text input                                                                                                      |
+| `min_length`  | number (optional)  | The minimum length of the text in the text input                                                                                 |
+| `max_length`  | number (optional)  | The maximum length of the text in the text input                                                                                 |
+| `required`    | boolean (optional) | Whether this text input is required                                                                                              |
+| `value`       | String (optional)  | The pre-filled value for the text input                                                                                          |
+| `placeholder` | String (optional)  | The placeholder text shown if the text input is empty                                                                            |
