+++ docs/parsables/embeds/embed-author.md
@@ -1,0 +1,9 @@
+`embed_author`
+
+Can also be parsed directly from a [User](/docs/values.md#User) value, or a string (which will only set `name`).
+
+| Value  | Type                                                                                                | Description                                                              |
+|--------|-----------------------------------------------------------------------------------------------------|--------------------------------------------------------------------------|
+| `name` | String                                                                                              | The displayed name of the author                                         |
+| `url`  | String (optional)                                                                                   | The URL link when clicking on the author name                            |
+| `icon` | String or Image from [Scarpet Graphics](https://github.com/replaceitem/scarpet-graphics) (optional) | The URL/File path/Image of the icon/avatar shown next to the author name |
