+++ docs/parsables/embeds/embed-field.md
@@ -1,0 +1,7 @@
+`embed_field`
+
+| Value    | Type               | Description                            |
+|----------|--------------------|----------------------------------------|
+| `name`   | String             | The name or title of this field        |
+| `value`  | String             | The value or description of this field |
+| `inline` | boolean (optional) | Whether this field is inline or not    |
