+++ docs/parsables/embeds/embed-footer.md
@@ -1,0 +1,6 @@
+`embed_footer`
+
+| Value  | Type                                                                                                | Description                                          |
+|--------|-----------------------------------------------------------------------------------------------------|------------------------------------------------------|
+| `text` | String                                                                                              | The footer text                                      |
+| `icon` | String or Image from [Scarpet Graphics](https://github.com/replaceitem/scarpet-graphics) (optional) | The URL/File path/image of the icon next to the text |
