+++ docs/parsables/embeds/embed.md
@@ -1,0 +1,14 @@
+`embed`
+
+| Value         | Type                                                                                                | Description                                                             |
+|---------------|-----------------------------------------------------------------------------------------------------|-------------------------------------------------------------------------|
+| `title`       | String                                                                                              | The title of the embed                                                  |
+| `url`         | String (optional)                                                                                   | The URL redirect when clicking on the embed title                       |
+| `description` | String (optional)                                                                                   | Description text below the title                                        |
+| `author`      | [Embed author](#Embed-author) (optional)                                                            | The author shown on top of the embed                                    |
+| `fields`      | List of [Embed fields](#Embed-field) (optional)                                                     | All fields inside the embed                                             |
+| `color`       | Color (optional)                                                                                    | The color of the embed                                                  |
+| `footer`      | [Embed footer](#Embed-footer) (optional)                                                            | The footer shown at the bottom of the embed                             |
+| `image`       | String or Image from [Scarpet Graphics](https://github.com/replaceitem/scarpet-graphics) (optional) | The URL/File path/image which will be shown in the embed                |
+| `thumbnail`   | String or Image from [Scarpet Graphics](https://github.com/replaceitem/scarpet-graphics) (optional) | The URL/File path/image which will be shown as a thumbnail in the embed |
+| `timestamp`   | [Timestamp](#Timestamp) (optional)                                                                  | The timestamp of the embed, which will be shown at the bottom           |
