+++ docs/parsables/timestamp.md
@@ -1,0 +1,7 @@
+`timestamp`
+
+Can also be parsed directly from a number, or a string `'now'`, which will parse to the current instant.
+
+| Value          | Type   | Description                       |
+|----------------|--------|-----------------------------------|
+| `epoch_millis` | number | Milliseconds since the epoch time |
