+++ docs/parsables/webhooks/webhook-message-profile.md
@@ -1,0 +1,8 @@
+`webhook_message_profile`
+
+Used when updating the webhook profile along with sending a message
+
+| Value    | Type              | Description                              |
+|----------|-------------------|------------------------------------------|
+| `name`   | String (optional) | The username of the webhook              |
+| `avatar` | String (optional) | A URL to the avatar shown on the webhook |
