+++ src/main/java/Discarpet/script/parsable/parsables/commands/MessageContextMenuBuilderParsable.java
@@ -1,4 +1,4 @@
-package Discarpet.script.parsable.parsables;
+package Discarpet.script.parsable.parsables.commands;
 
 import Discarpet.script.parsable.ParsableClass;
 import Discarpet.script.parsable.ParsableConstructor;
