+++ src/main/java/Discarpet/script/parsable/parsables/components/SelectMenuParsable.java
@@ -1,4 +1,4 @@
-package Discarpet.script.parsable.parsables;
+package Discarpet.script.parsable.parsables.components;
 
 import Discarpet.script.parsable.Optional;
 import Discarpet.script.parsable.ParsableClass;
