+++ src/main/java/Discarpet/script/util/Registration.java
@@ -10,27 +10,27 @@ import Discarpet.script.functions.ValueFromId;
 import Discarpet.script.parsable.Parser;
 import Discarpet.script.parsable.parsables.AllowedMentionsParsable;
 import Discarpet.script.parsable.parsables.AttachmentParsable;
-import Discarpet.script.parsable.parsables.ButtonParsable;
+import Discarpet.script.parsable.parsables.components.ButtonParsable;
 import Discarpet.script.parsable.parsables.ColorParsable;
-import Discarpet.script.parsable.common.ComponentParsable;
-import Discarpet.script.parsable.parsables.EmbedAuthorParsable;
-import Discarpet.script.parsable.parsables.EmbedFieldParsable;
-import Discarpet.script.parsable.parsables.EmbedFooterParsable;
-import Discarpet.script.parsable.parsables.EmbedParsable;
+import Discarpet.script.parsable.parsables.components.ComponentParsable;
+import Discarpet.script.parsable.parsables.embeds.EmbedAuthorParsable;
+import Discarpet.script.parsable.parsables.embeds.EmbedFieldParsable;
+import Discarpet.script.parsable.parsables.embeds.EmbedFooterParsable;
+import Discarpet.script.parsable.parsables.embeds.EmbedParsable;
 import Discarpet.script.parsable.parsables.commands.MessageContextMenuBuilderParsable;
 import Discarpet.script.parsable.parsables.TimestampParsable;
 import Discarpet.script.parsable.parsables.MessageContentParsable;
-import Discarpet.script.parsable.parsables.ModalParsable;
-import Discarpet.script.parsable.parsables.SelectMenuOptionParsable;
-import Discarpet.script.parsable.parsables.SelectMenuParsable;
+import Discarpet.script.parsable.parsables.components.ModalParsable;
+import Discarpet.script.parsable.parsables.components.SelectMenuOptionParsable;
+import Discarpet.script.parsable.parsables.components.SelectMenuParsable;
 import Discarpet.script.parsable.parsables.commands.SlashCommandBuilderParsable;
 import Discarpet.script.parsable.parsables.commands.SlashCommandOptionChoiceParsable;
 import Discarpet.script.parsable.parsables.commands.SlashCommandOptionParsable;
-import Discarpet.script.parsable.parsables.TextInputParsable;
+import Discarpet.script.parsable.parsables.components.TextInputParsable;
 import Discarpet.script.parsable.parsables.commands.UserContextMenuBuilderParsable;
-import Discarpet.script.parsable.parsables.WebhookMessageProfileParsable;
-import Discarpet.script.parsable.parsables.WebhookProfileParsable;
-import Discarpet.script.parsable.parsables.WebhookProfileUpdaterParsable;
+import Discarpet.script.parsable.parsables.webhooks.WebhookMessageProfileParsable;
+import Discarpet.script.parsable.parsables.webhooks.WebhookProfileParsable;
+import Discarpet.script.parsable.parsables.webhooks.WebhookProfileUpdaterParsable;
 import Discarpet.script.values.AttachmentValue;
 import Discarpet.script.values.commands.MessageContextMenuValue;
 import Discarpet.script.values.commands.SlashCommandValue;
